requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
chainlit>=1.0.0
Pillow>=9.0.0
google-generativeai>=0.3.0
python-dotenv
openpyxl